Support other network links
